fixed bugs after strict types

+++ src/DND/Domain/Race/Race.php
@@ -42,7 +42,7 @@ class Race
         return $this->speed;
     }
 
-    public function getNightvision(): int
+    public function getNightvision(): float
     {
         return $this->nightvision;
     }
